Update core
- fix error message "DAGs not found" during download of updated miners

+++ Modules/Include.psm1
@@ -2825,13 +2825,15 @@ function Expand-WebRequest {
                 $Rm_Paths = @("DAGs")
                 $Rm_Paths | Foreach-Object {
                     $Rm_Path = Join-Path $Path_Bak $_
-                    try {
-                        Get-ChildItem $Rm_Path -File | Foreach-Object {
-                            Remove-Item $_.FullName -Force
+                    if (Test-Path $Rm_Path) {
+                        try {
+                            Get-ChildItem $Rm_Path -File | Foreach-Object {
+                                Remove-Item $_.FullName -Force
+                            }
+                        } catch {
+                            if ($Error.Count){$Error.RemoveAt(0)}
+                            Write-Log -Level Warn "Downloader: Could not remove from backup path $_. Please do this manually, root might be needed ($($_.Exception.Message))"
                         }
-                    } catch {
-                        if ($Error.Count){$Error.RemoveAt(0)}
-                        Write-Log -Level Warn "Downloader: Could not to remove from backup path $_. Please do this manually, root might be needed ($($_.Exception.Message))"
                     }
                 }
                 $SkipBackups = if ($EnableMinerBackups) {3} else {0}
